+++ modules/api_gw/variables.tf
@@ -4,21 +4,14 @@ variable "name_prefix" {
   type = string
 }
 
-
 variable "ENV" {
   type = string
-  
-}
 
+}
 variable "aws_lambda_function_arn" {
   type = string
 }
 
 variable "log_group_arn" {
-  type = string
-  
-}
-
-variable "MODE" {
   type = string
 }
