+++ modules/lambda/main.tf
@@ -2,10 +2,10 @@
 # lambda
 #=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-
 resource "aws_lambda_function" "app" {
-  function_name = "${var.name_prefix}-${var.ENV}-${var.MODE}"
+  function_name = "${var.name_prefix}-${var.ENV}"
   role          = var.lambda_role_arn
   runtime       = "python3.12"
-  handler       = "function-${var.ENV}-${var.MODE}.lambda_handler"
-  filename         = "../../src/lambda/function-${var.ENV}-${var.MODE}.zip"
-  source_code_hash = filebase64sha256("../../src/lambda/function-${var.ENV}-${var.MODE}.zip")
+  handler       = "function-${var.ENV}.lambda_handler"
+  filename         = "../../src/lambda/function-${var.ENV}.zip"
+  source_code_hash = filebase64sha256("../../src/lambda/function-${var.ENV}.zip")
 }
